+++ Adj/Algebra.hs
@@ -1,0 +1,1 @@
+module Adj.Algebra where