+++ Adj/Program.hs
@@ -1,0 +1,1 @@
+module Adj.Program where