+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # 0.0.1
+* Define `Adj.Algebra` module
+* Define `Adj.Program` module